Move tables to subdir
