+++ frontend/src/api/stats.ts
@@ -1,4 +1,4 @@
-import { LazyResult } from '../component/LazyTableSimple';
+import { LazyResult } from '../component/table/LazyTableSimple';
 import { apiCall } from './common';
 
 export interface DatabaseStats {
