+++ frontend/src/component/formik/AutoSubmitPaginationField.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import React from 'react';
 import { useFormikContext } from 'formik';
 import { logErr } from '../../util/errors';
-import { RowsPerPage } from '../LazyTable';
+import { RowsPerPage } from '../table/LazyTable';
 
 export const AutoSubmitPaginationField = ({
     page,
