+++ frontend/src/component/table/BanGroupTable.tsx
@@ -12,20 +12,24 @@ import Grid from '@mui/material/Unstable_Grid2';
 import { formatDuration, intervalToDuration } from 'date-fns';
 import { Formik } from 'formik';
 import * as yup from 'yup';
-import { apiGetBansGroups, BanGroupQueryFilter, GroupBanRecord } from '../api';
-import { useUserFlashCtx } from '../contexts/UserFlashCtx';
-import { logErr } from '../util/errors';
-import { renderDate } from '../util/text';
+import {
+    apiGetBansGroups,
+    BanGroupQueryFilter,
+    GroupBanRecord
+} from '../../api';
+import { useUserFlashCtx } from '../../contexts/UserFlashCtx';
+import { logErr } from '../../util/errors';
+import { renderDate } from '../../util/text';
+import { DeletedField, deletedValidator } from '../formik/DeletedField';
+import { FilterButtons } from '../formik/FilterButtons';
+import { GroupIdField, groupIdFieldValidator } from '../formik/GroupIdField';
+import { SourceIdField, sourceIdValidator } from '../formik/SourceIdField';
+import { SteamIDSelectField } from '../formik/SteamIDSelectField';
+import { TargetIDField, targetIdValidator } from '../formik/TargetIdField';
+import { ModalBanGroup, ModalUnbanGroup } from '../modal';
+import { BanGroupModalProps } from '../modal/BanGroupModal';
 import { LazyTable, Order, RowsPerPage } from './LazyTable';
 import { TableCellBool } from './TableCellBool';
-import { DeletedField, deletedValidator } from './formik/DeletedField';
-import { FilterButtons } from './formik/FilterButtons';
-import { GroupIdField, groupIdFieldValidator } from './formik/GroupIdField';
-import { SourceIdField, sourceIdValidator } from './formik/SourceIdField';
-import { SteamIDSelectField } from './formik/SteamIDSelectField';
-import { TargetIDField, targetIdValidator } from './formik/TargetIdField';
-import { ModalBanGroup, ModalUnbanGroup } from './modal';
-import { BanGroupModalProps } from './modal/BanGroupModal';
 
 interface GroupBanFilterValues {
     group_id: string;
