+++ frontend/src/component/table/TableCellBool.tsx
