+++ frontend/src/page/AdminReportsPage.tsx
@@ -15,9 +15,7 @@ import {
     ReportWithAuthor
 } from '../api';
 import { ContainerWithHeader } from '../component/ContainerWithHeader';
-import { LazyTable, Order, RowsPerPage } from '../component/LazyTable';
 import { LoadingIcon } from '../component/LoadingIcon';
-import { TableCellLink } from '../component/TableCellLink';
 import { FilterButtons } from '../component/formik/FilterButtons';
 import {
     ReportStatusField,
@@ -32,6 +30,8 @@ import {
     TargetIDField,
     targetIdValidator
 } from '../component/formik/TargetIdField';
+import { LazyTable, Order, RowsPerPage } from '../component/table/LazyTable';
+import { TableCellLink } from '../component/table/TableCellLink';
 import { logErr } from '../util/errors';
 import { renderDateTime } from '../util/text';
