+++ frontend/src/page/ForumPage.tsx
@@ -16,8 +16,8 @@ import { apiForum, apiGetThreads, Forum, ForumThread } from '../api/forum';
 import { ContainerWithHeaderAndButtons } from '../component/ContainerWithHeaderAndButtons';
 import { ForumRowLink } from '../component/ForumRowLink';
 import { VCenteredElement } from '../component/Heading';
-import { RowsPerPage } from '../component/LazyTable';
 import { ModalForumThreadEditor } from '../component/modal';
+import { RowsPerPage } from '../component/table/LazyTable';
 import { useCurrentUserCtx } from '../contexts/CurrentUserCtx';
 import { useUserFlashCtx } from '../contexts/UserFlashCtx';
 import { logErr } from '../util/errors';
