+++ frontend/src/page/STVPage.tsx
@@ -1,7 +1,7 @@
 import React, { JSX } from 'react';
 import VideocamIcon from '@mui/icons-material/Videocam';
 import { ContainerWithHeader } from '../component/ContainerWithHeader';
-import { STVTable } from '../component/STVTable';
+import { STVTable } from '../component/table/STVTable';
 
 export const STVPage = (): JSX.Element => {
     return (
