+++ frontend/src/page/StatsWeaponOverallPage.tsx
@@ -8,10 +8,10 @@ import {
     Weapon
 } from '../api';
 import { ContainerWithHeader } from '../component/ContainerWithHeader';
-import { RowsPerPage } from '../component/LazyTable';
-import { LazyTableSimple } from '../component/LazyTableSimple';
 import { PersonCell } from '../component/PersonCell';
 import { fmtWhenGt } from '../component/PlayersOverallContainer';
+import { RowsPerPage } from '../component/table/LazyTable';
+import { LazyTableSimple } from '../component/table/LazyTableSimple';
 import { defaultFloatFmtPct, humanCount } from '../util/text';
 
 interface WeaponStatsContainerProps {
